Moving files to test write
